Compiler warnings (unknown pragmas) SOA files removed from lib-projects
git-svn-id: https://svn.code.sf.net/p/tcnopen/trdp/trunk@2207 3b5a3598-5f4e-4449-9e63-bd40438bfec0

+++ trdp/test/localtest/api_test_2.c
@@ -3589,7 +3589,7 @@ int main (int argc, char *argv[])
             case 'v':   /*  version */
                 printf("%s: Version %s\t(%s - %s)\n",
                        argv[0], APP_VERSION, __DATE__, __TIME__);
-                printf("No. of tests: %lu\n", sizeof(testArray) / sizeof(void *) - 2);
+                printf("No. of tests: %lu\n", (unsigned long) (sizeof(testArray) / sizeof(void *) - 2));
                 exit(0);
                 break;
             case 'h':
